Bring parity with EDMC 5.3.1
