+++ python/requirements-dev.txt
@@ -1,0 +1,51 @@
+# So that you don't get warnings like:
+# Using legacy 'setup.py install' for flake8-annotations-coverage, since package 'wheel' is not installed.
+wheel
+
+# We can't rely on just picking this up from either the base (not venv),
+# or venv-init-time version.  Specify here so that dependabot will prod us
+# about new versions.
+setuptools==60.9.3
+
+# Static analysis tools
+flake8==4.0.1
+flake8-annotations-coverage==0.0.5
+flake8-cognitive-complexity==0.1.0
+flake8-comprehensions==3.8.0
+flake8-docstrings==1.6.0
+isort==5.10.1
+flake8-isort==4.1.1
+flake8-json==21.7.0
+flake8-noqa==1.2.1
+flake8-polyfill==1.0.2
+flake8-use-fstring==1.3
+
+mypy==0.931
+pep8-naming==0.12.1
+safety==1.10.3
+types-requests==2.27.10
+
+# Code formatting tools
+autopep8==1.6.0
+
+# HTML changelogs
+grip==4.6.0
+
+# Packaging
+# Used to put together a WiX configuration from template/auto-gen
+lxml==4.8.0
+# We only need py2exe on windows.
+# Pre-release version addressing semantic_version 2.9.0+ issues:
+# <https://github.com/py2exe/py2exe/issues/126>
+py2exe==0.11.1.0; sys_platform == 'win32'
+
+# Testing
+pytest==7.0.1
+pytest-cov==3.0.0  # Pytest code coverage support
+coverage[toml]==6.3.1 # pytest-cov dep. This is here to ensure that it includes TOML support for pyproject.toml configs
+# For manipulating folder permissions and the like.
+pywin32==303; sys_platform == 'win32'
+
+
+# All of the normal requirements
+-r requirements.txt
