+++ python/requirements.txt
@@ -1,0 +1,11 @@
+certifi==2021.10.8
+requests==2.27.1
+watchdog==2.1.6
+# Commented out because this doesn't package well with py2exe
+infi.systray==0.1.12; sys_platform == 'win32'
+# argh==0.26.2 watchdog dep
+# pyyaml==5.3.1 watchdog dep
+semantic-version==2.9.0
+
+# Base requirement for MacOS
+pyobjc; sys_platform == 'darwin'
