+++ python/ttkHyperlinkLabel.py
@@ -1,14 +1,12 @@
+import sys
 import tkinter as tk
 import webbrowser
-from sys import platform
 from tkinter import font as tkFont
 from tkinter import ttk
 
-if platform == 'win32':
+if sys.platform == 'win32':
     import subprocess
-    from winreg import (
-        HKEY_CLASSES_ROOT, HKEY_CURRENT_USER, CloseKey, OpenKeyEx, QueryValueEx
-    )
+    from winreg import HKEY_CLASSES_ROOT, HKEY_CURRENT_USER, CloseKey, OpenKeyEx, QueryValueEx
 
 # A clickable ttk Label
 #
@@ -18,19 +16,22 @@ if platform == 'win32':
 #   popup_copy: Whether right-click on non-empty label text pops up a context menu with a 'Copy' option. Defaults to no context menu. If popup_copy is a function it will be called with the current label text and should return a boolean.
 #
 # May be imported by plugins
-class HyperlinkLabel(platform == 'darwin' and tk.Label or ttk.Label, object):
+
+
+class HyperlinkLabel(sys.platform == 'darwin' and tk.Label or ttk.Label, object):
 
     def __init__(self, master=None, **kw):
         self.url = 'url' in kw and kw.pop('url') or None
         self.popup_copy = kw.pop('popup_copy', False)
-        self.underline = kw.pop('underline', None)	# override ttk.Label's underline
+        self.underline = kw.pop('underline', None)  # override ttk.Label's underline
         self.foreground = kw.get('foreground') or 'blue'
-        self.disabledforeground = kw.pop('disabledforeground', ttk.Style().lookup('TLabel', 'foreground', ('disabled',)))	# ttk.Label doesn't support disabledforeground option
+        self.disabledforeground = kw.pop('disabledforeground', ttk.Style().lookup(
+            'TLabel', 'foreground', ('disabled',)))  # ttk.Label doesn't support disabledforeground option
 
-        if platform == 'darwin':
+        if sys.platform == 'darwin':
             # Use tk.Label 'cos can't set ttk.Label background - http://www.tkdocs.com/tutorial/styles.html#whydifficult
             kw['background'] = kw.pop('background', 'systemDialogBackgroundActive')
-            kw['anchor'] = kw.pop('anchor', tk.W)	# like ttk.Label
+            kw['anchor'] = kw.pop('anchor', tk.W)  # like ttk.Label
             tk.Label.__init__(self, master, **kw)
         else:
             ttk.Label.__init__(self, master, **kw)
@@ -39,16 +40,16 @@ class HyperlinkLabel(platform == 'darwin' and tk.Label or ttk.Label, object):
 
         self.menu = tk.Menu(None, tearoff=tk.FALSE)
         # LANG: Label for 'Copy' as in 'Copy and Paste'
-        self.menu.add_command(label=_('Copy'), command = self.copy)  # As in Copy and Paste
-        self.bind(platform == 'darwin' and '<Button-2>' or '<Button-3>', self._contextmenu)
+        self.menu.add_command(label=_('Copy'), command=self.copy)  # As in Copy and Paste
+        self.bind(sys.platform == 'darwin' and '<Button-2>' or '<Button-3>', self._contextmenu)
 
         self.bind('<Enter>', self._enter)
         self.bind('<Leave>', self._leave)
 
         # set up initial appearance
-        self.configure(state = kw.get('state', tk.NORMAL),
-                       text = kw.get('text'),
-                       font = kw.get('font', ttk.Style().lookup('TLabel', 'font')))
+        self.configure(state=kw.get('state', tk.NORMAL),
+                       text=kw.get('text'),
+                       font=kw.get('font', ttk.Style().lookup('TLabel', 'font')))
 
     # Change cursor and appearance depending on state and text
     def configure(self, cnf=None, **kw):
@@ -70,17 +71,18 @@ class HyperlinkLabel(platform == 'darwin' and tk.Label or ttk.Label, object):
 
         if 'font' in kw:
             self.font_n = kw['font']
-            self.font_u = tkFont.Font(font = self.font_n)
-            self.font_u.configure(underline = True)
+            self.font_u = tkFont.Font(font=self.font_n)
+            self.font_u.configure(underline=True)
             kw['font'] = self.underline is True and self.font_u or self.font_n
 
         if 'cursor' not in kw:
             if (kw['state'] if 'state' in kw else str(self['state'])) == tk.DISABLED:
-                kw['cursor'] = 'arrow'	# System default
+                kw['cursor'] = 'arrow'  # System default
             elif self.url and (kw['text'] if 'text' in kw else self['text']):
-                kw['cursor'] = platform=='darwin' and 'pointinghand' or 'hand2'
+                kw['cursor'] = sys.platform == 'darwin' and 'pointinghand' or 'hand2'
             else:
-                kw['cursor'] = (platform=='darwin' and 'notallowed') or (platform=='win32' and 'no') or 'circle'
+                kw['cursor'] = (sys.platform == 'darwin' and 'notallowed') or (
+                    sys.platform == 'win32' and 'no') or 'circle'
 
         super(HyperlinkLabel, self).configure(cnf, **kw)
 
@@ -89,22 +91,22 @@ class HyperlinkLabel(platform == 'darwin' and tk.Label or ttk.Label, object):
 
     def _enter(self, event):
         if self.url and self.underline is not False and str(self['state']) != tk.DISABLED:
-            super(HyperlinkLabel, self).configure(font = self.font_u)
+            super(HyperlinkLabel, self).configure(font=self.font_u)
 
     def _leave(self, event):
         if not self.underline:
-            super(HyperlinkLabel, self).configure(font = self.font_n)
+            super(HyperlinkLabel, self).configure(font=self.font_n)
 
     def _click(self, event):
         if self.url and self['text'] and str(self['state']) != tk.DISABLED:
             url = self.url(self['text']) if callable(self.url) else self.url
             if url:
-                self._leave(event)	# Remove underline before we change window to browser
+                self._leave(event)  # Remove underline before we change window to browser
                 openurl(url)
 
     def _contextmenu(self, event):
         if self['text'] and (self.popup_copy(self['text']) if callable(self.popup_copy) else self.popup_copy):
-            self.menu.post(platform == 'darwin' and event.x_root + 1 or event.x_root, event.y_root)
+            self.menu.post(sys.platform == 'darwin' and event.x_root + 1 or event.x_root, event.y_root)
 
     def copy(self):
         self.clipboard_clear()
@@ -112,13 +114,14 @@ class HyperlinkLabel(platform == 'darwin' and tk.Label or ttk.Label, object):
 
 
 def openurl(url):
-    if platform == 'win32':
+    if sys.platform == 'win32':
         # On Windows webbrowser.open calls os.startfile which calls ShellExecute which can't handle long arguments,
         # so discover and launch the browser directly.
         # https://blogs.msdn.microsoft.com/oldnewthing/20031210-00/?p=41553
 
         try:
-            hkey = OpenKeyEx(HKEY_CURRENT_USER, r'Software\Microsoft\Windows\Shell\Associations\UrlAssociations\https\UserChoice')
+            hkey = OpenKeyEx(HKEY_CURRENT_USER,
+                             r'Software\Microsoft\Windows\Shell\Associations\UrlAssociations\https\UserChoice')
             (value, typ) = QueryValueEx(hkey, 'ProgId')
             CloseKey(hkey)
             if value in ['IE.HTTP', 'AppXq0fevzme2pys62n3e0fbqa7peapykr8v']:
@@ -128,7 +131,7 @@ def openurl(url):
             else:
                 cls = value
         except:
-            cls  = 'https'
+            cls = 'https'
 
         if cls:
             try:
